// src/app/interfaces/country.interface.ts
export interface Country {
  id: string;
  name: string;
  nameSE: string;
  nationality: string;
  nationalitySEName: string;
}

export interface CreateCountry {
  id: string;
  name: string;
  nameSE: string;
  nationality: string;
  nationalitySEName: string;
}

export interface EditCountry {
  id: string;
  name: string;
  nameSE: string;
  nationality: string;
  nationalitySEName: string;
}